make new controller and router for the crew endpoint
