+++ packages/server/api/controllers/movies.controller.js
@@ -41,28 +41,6 @@ const getMoviesByCategory = async (categoryId) => {
   }
 };
 
-const getCrewOfMovieByID = async (id) => {
-  if (!id) {
-    throw new HttpError('ID should be a number', 400);
-  }
-  try {
-    const crew = await knex
-      .select(
-        'crew_members.id',
-        'full_name',
-        'birth_date',
-        'image_location',
-        'role',
-      )
-      .from('movie_crew')
-      .join('crew_members', 'movie_crew.crew_member_id', '=', 'crew_members.id')
-      .where('movie_id', id);
-    return crew;
-  } catch (error) {
-    throw new HttpError(error.message, 500);
-  }
-};
-
 const editMovie = async (movieID, updateMovie) => {
   if (!movieID) {
     throw new HttpError('movie ID should be a number', 400);
@@ -148,7 +126,6 @@ module.exports = {
   editMovie,
   deleteMovie,
   createMovie,
-  getCrewOfMovieByID,
   getMoviesByCategory,
   getMovieList,
   getFeaturedMovie,
