+++ packages/server/api/routes/index.js
@@ -3,6 +3,7 @@ const express = require('express');
 const router = express.Router();
 
 const movies = require('./movies.router');
+const crew = require('./crew.router');
 
 const swaggerJsDoc = require('swagger-jsdoc');
 const swaggerUi = require('swagger-ui-express');
@@ -28,5 +29,6 @@ const swaggerDocument = swaggerJsDoc(swaggerOptions);
 router.use('/documentation', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
 
 router.use('/movies', movies);
+router.use('/crew', crew);
 
 module.exports = router;
